relax version 1.3.0.
This is the first release of the 1.3 development line.  In addition to all the capabilities of relax
1.2.10 the new features of this release include the addition of a unit test framework, preliminary
support for PyMOL, the creation of PDB files, the ability to display the diffusion tensor encoded as
a PDB file using PyMOL, the ability to display the distribution of XH bond vectors used in the
analysis via a PDB file, the new user function class 'structure' containing the old user function
'pdb()' which has been split and redesigned, removal of apply(func, args) calls from the
minimisation code, improved spin system selection using Boolean operators, and a few other small
features, bugfixes, and program changes.  As the user function names have been and will continue to
be modified please do not expect either the backwards or forwards compatibility of relax scripts
written for the 1.3 relax versions.