Added an import statement to the 'data/__init__.py' file to simplify the module.
The class Data() can now be accessed by typing 'from data import Data' rather than
'from data.main import Data'.  This will simplify the usage of this module for the data model
redesign and allow file names in 'data/' to be changed without affecting the operation of the rest
of relax.