Completion of the core of the molecule-residue-spin data structures.
All of the initialisation of the class variables and objects in all six classes have been shifted
into __init__() methods so that new copies of the objects are created when molecules, residues, or
spins are added to the structure.

The __repr__() methods of all classes have been updated and return a proper text representation of
the objects.

The XList class docstrings have been updated.