Bug fix for the prototype copying.
These changes cause the 'test_copy' and 'test_copy_current' unit tests of 'test_pipe.py' to pass.

The prototype '__copy__()' method has been renamed to '__clone__()' as '__copy__()' is reserved as a
replacement method for the 'copy' method of the 'copy' module.  As the 'deepcopy' method was not
copying all of the objects of the pipe container correctly, the '__deepcopy__() replacement method
has been added to the prototype base class.

All subclassing of 'ListType' and 'DictType' have been replaced by 'list' and 'dict'.