Fix for bug #8996 (https://gna.org/bugs/?8996).
This bug was reported by Carl Diehl.

The issue was that the data structure 'data.heteronuc' did not exist and is now called
'pdb_heteronuc'.  This issue had been fixed in the 1.3 line but was incomplete in the 1.2 line (the
printing of values was ok but the errors and simulations were incorrect).