Fix for bug #8953 (https://gna.org/bugs/?8953).
The bug was introduced by a fix to the model-free results writing function.  In relax versions 1.2.9
and below, no data about a residue was being stored in the results file if it was deselected.  This
behaviour undesirable as a spin system could be deselected temporarily and later on reselected.  If
this process is deliberately interrupted by using 'results.write()' followed by 'results.read()' to
resume, then there will be data loss.

The fix was simple: check for the presence of the 'data.num_ri' data structure.