Improvements to the 'full_analysis.py' sample script.
The script's documentation has been shifted from the __init__() method docstring to the module
docstring.

All of the options which need to be modified by the user have been shifted to the top of the script
as global variables.  This means that the user doesn't even need to look at the complex code to
operate the script.