Fix for bug #9389 (https://gna.org/bugs/?9389).
This bug was caused by the move away from using the __builtin__ namespace.  As the file is now
treated as a module from which errors must be imported, the AllRelaxErrors object was not being
created.  The solution was to add the function 'all_errors()' which returns a list of all the
RelaxError objects.