Fix for bug #10320 (https://gna.org/bugs/?10320).
This bug was reported by Douglas Kojetin <douglas dot kojetin at gmail dot com>.

The bug was introduced at https://mail.gna.org/public/relax-commits/2007-05/msg00016.html in r3294.
When ported to the 1.3 line, the bug somehow did not make it into that code.  See
https://mail.gna.org/public/relax-commits/2007-06/msg00006.html.

The bug is that sometimes an the residue specific parameter array is set to [] rather than there
being no parameter array.  This situation was trying to be caught in the 1.2 code, however there was
a dumb bug - the closing bracket was placed in the wrong position!