prompt.interpreter.interact_script() has been modified to allow RelaxErrors to be raised.
Previously running a script would catch the error, send a nice message to STDERR, and quit.  Now
the function can raise the error if asked.  This is required for the system tests (because of the
unittest framework).  Alternatively the status of the script could have been used by the test, but
this is easier for the test writer and makes the code more flexible.