Fix for bug #8997 (https://gna.org/bugs/index.php?8997).
This bug was reported by Carl Diehl (carl dot diehl at bpc dot lu dot se).

The problem was because of a change between relax versions 1.0 and 1.2.  Previously the residue
number was being searched for, but the changes used the absolute position to significantly speed up
the code.  But the absolute positions are different in a Modelfree 'mfout' file if simulations are
run.  The absolute position in this situation is now properly calculated and the test data attached
to the bug report by Seb now works.