Updated the generic_fns.value module functions set() and partition_params() to be more general.
The updating is not complete, but works for most of the current codebase.  The assumption of a data
analysis type having a diffusion or other tensor has been removed.  This has been replaced by the
concept that the params are now partitioned into those that are spin specific and those that are
not.  The diffusion tensor stuff will be moved into the model-free code.