Bug fix for the generic_fns.value.set() function.
The differentiation between a spin specific parameter and a global parameter was not working for
specific analyses with no spin specific parameters.  The return_data_name() function was being used
to differentiate, which makes no sense.  So instead the is_spin_param() function has been introduced
to allow the specific code to answer the question explicitly.  Note that this will temporarily break
many things.