Fix for bug #11378 (https://gna.org/bugs/?11378).
This bug was reported by Ryan Hoffman (https://gna.org/users/rydog).

The fix for the 1.3 line is very different to that for the 1.2 line.  These functions have changed
significantly, and the pdb() user function doesn't exist any more!  The concept is similar though.
The xh_vect variable is now set to None rather than being non-existant, and the spin is deselected.