Modifications to the structure.read_pdb() user function.
The 'parser' argument has been added to allow different PDB parsers to be selected in the future.
To allow for the new relax design, the docstring has been updated with respect to the loading of the
sequence and the 'spin_id' argument has been added to allow the selection of specific atoms of the
PDB file to be placed as SpinContainers in the relax data store.