Split up of the generic_fns.selection module.
All of the methods relating to the molecules, residues, and spins (including the spin selection
string and associated functions and the spin selection object) have been shifted into the
generic_fns.mol_res_spin module.  The functions remaining in generic_fns.selection are related to
purely the selection and deselection of spins.