Changed the design of the relax data store singleton.
The singleton saving and loading problem is discussed in the thread starting at:
https://mail.gna.org/public/relax-devel/2007-11/msg00021.html (Message-id:
<7f080ed10711201001s59a00533pf3c0ee3f88a79edb@mail.gmail.com>).  For more history, see the thread
starting at https://mail.gna.org/public/relax-devel/2007-03/msg00046.html (Message-id:
<7f080ed10703140609p70d35295ye10ee5939da494a2@mail.gmail.com>) and the original singleton idea at
https://mail.gna.org/public/relax-devel/2007-03/msg00012.html (Message-id:
<7f080ed10703070006s309a1e1do38e621f48d8f6d9e@mail.gmail.com>).

This reverts to the original singleton behaviour in the last link.  Note that this will be reverted
to the current singleton design once a good solution to the pickling issues can be found.