Bug fix for the Unit_test_runner.
The get_module_relative_path() function was sometimes failing because of the format of the
package_path argument.  This should be a string but, when run from an outside directory, often this
path would already be split up into the segment path list.  The fix just catches this so that the
function doesn't try to segment the segmented list.