Bug fixes and expansion of the __linear_constraints() method.
The bug fix is that the b array was not being scaled!  This is essential when diagonal scaling is
active.

The added feature is that the pN weight of state N, which is not a parameter of the model, is now
constrained to be between 0 and 1.  This is because the value is equal to 1 minus the sum of all
other probabilities, hence linear constraints can be created.