Creation of a branch for the clean up of references to the current data pipe.
Rather than having 'ds.current_pipe' and 'ds[ds.current_pipe]' hard coded throughout relax, the aim
of this branch is to replace these with the API functions of the generic_fns.pipes module.  This
will allow the relax data store representation of the current pipe to evolve without hurting the
rest of relax.