Fix for a bug introduced at r7519.
The problem was being picked up by the unit tests.  The duplicate_data() model-free method was
failing when sequence data was missing from the source data pipe.  This was supposed to be caught
in the code before r7519 but the wrong data pipe was being checked and the code was removed at that
revision.