Added more for the reading of intensities from a nmrview peak list.
There is still a need to fix the reading of the assignments which are of the format number.nucleus
(e.g. 22.HN).

Moreover, a functionality to chose whether the intensity or volume column is read could be added as
proposed in a post at: https://mail.gna.org/public/relax-devel/2008-10/msg00026.html
(Message-id: <48EE7028.6060409@ulaval.ca>).