Important bug fix!!
The theta angle of the spheroidal diffusion tensor was being improperly wrapped between 0 and pi.
This was causing the glide reflection symmetry to be broken, hence the tensor was incorrect.  The
result would be longer optimisation times for the spheroid tensors and more iterations than
necessary in the full_analysis.py script.