Added more to the XEasy support test.
Now, the test fails when reaching the first residue with a non HN/N assignment. In fact, the reading
of this residue is fine, but no intensity is extracted... hence the failure of the test.

Now, what do we want to do ? Do we want to support side chain assignments mixed with backbone
assignments ? In other words, do we want to support multiuple assignment types in a single
analysis ?