Fixes for the model-free LaTeX table generation sample script.
These fixes do not affect the system test result yet - as LaTeX compilation will not be part of the
test suite - but fix the LaTeX compilation.  These will be ported back to the sample script.

The docstring has been modified as the booktabs package is needed for this table.  The '&'
characters in the spin_id string have been escaped to avoid this being interpreted as a column
separator.  The 'Residue' column has been renamed to 'Spin'.  And finally missing model-free
parameter errors can be handled.