Fix for bug #12487 (https://gna.org/bugs/index.php?12487).
This bug was reported by Tyler Reddy (TREDDY at dal dot ca).

The problem was that missing XH bond vector data was being manifested as the spin.xh_vect object
being set to None.  The overfit_deselect() function which is responsible for deselecting unsuitable
spins prior to optimisation was only checking if the object existed, not if it was None!