Fix for bug #12582 (https://gna.org/bugs/index.php?12582).
This was reported by Chris Brosey (chris dot a dot brosey att vanderbilt dot edu).

The problem was that the frq_labels data structure in the SpinContainer was sometimes set to []
rather than the data structure being missing.  Now the structure is being tested if it is either
None or [] before being accessed in the model-free return_units() method.