Final fix for bug #12591 (https://gna.org/bugs/index.php?12591).
This bug was reported by Chris Brosey (chris dot a dot brosey att vanderbilt dot edu).

The problem was that the global_index counter in the model-free model_loop() method was supposed to
be incremented at the start of the loop.  The deselected spin skipping code was located before this
though, causing the counter to not be incremented!