Continued to implement the reading of the generic peak intensity file.
The code now calculates the number of delays (and associated intensities) input.

The remaining tasks include sorting the delays and intensities, which is different from when other
peak intensity files are used (e.g. XEasy, Sparky, nmrview) since the delays are part of the input
file header (in the case of the generic file).