Partial fix for bug #12735 (https://gna.org/bugs/index.php?12735).
The problem fixed here is that of the missing molecule container.  This was caused by the weirdness
of the Scientific python PDB reader splitting one modified lactose molecule into 2 molecules.  The
lactose part was unnamed, while the modifier was named after the residues 'UNK'.  The
generic_fns.mol_res_spin.return_molecule() function was having problems with this concept of the
first molecule unnamed and the second named.  This is now fixed and the unnamed molecule can be
retrieved.