Another partial fix for bug #12735 (https://gna.org/bugs/index.php?12735).
This problem was trigged by the internal reader (where the problem lies) together with the
gromacs_phthalic_acid.pdb file.  The issue was that the __fill_object_from_pdb() was operating on
all structures currently loaded into relax.  Hence with the 2 models of this PDB file, the first
molecule had all the atoms of the first and second PDB models combined.