Fixes for the 2 test_atom_loop_spin_selection2() unit tests.
This problem was reported by Jack Howarth (https://gna.org/users/jhowarth) in the post at
https://mail.gna.org/public/relax-users/2008-10/msg00089.html (Message-id:
<20081024141150.GA879@bromo.msbb.uc.edu>).

The problem was with the position array comparison as this was being converted to a string and then
compared.  The string representations of the numpy arrays is obviously different on the intel vs.
powerpc chips.