Corrected the way the scaling matrix is assembled.
This is as proposed by Ed in a post at:
https://mail.gna.org/public/relax-devel/2009-01/msg00079.html
(Message-id: <7f080ed10901130637k34ce0fc1l57091178e09940a8@mail.gmail.com>)

The scaling values are now based on the default values for the different parameters which were
slightly modified.

The only parameter for which the average is still used (as for intensities in the 'relax_fit.py'
code) is 'R2eff'.