Fixed an annoying bug in the RelaxError debugging mode.
The problem here is that the relax state was being saved in the __init__() methods if in debug mode.
The result was that in try statements which catch and allow for RelaxErrors, the relax state was
being saved even though the error message and exception was suppressed.  The fix was to shift the
state saving code into the __str__() method.

A side effect of this change is a simplification of all the RelaxError classes as the debug code is
no longer duplicated in each class.