The execfile statement has been replaced with a call to the exec(compile()) function.
This was done with the python 2.x to 3.x script with the commands:
2to3 -w -f exec relax .

The change is in preparation for Python 3.0, and relax still works with Python versions 2.4, 2.5,
and 2.6.  The syntax has changed from:

execfile(script_file, local)

to:

exec(compile(open(script_file).read(), script_file, 'exec'), local)

Hmmm, not much of a language simplification!