Significant improvements to the __find_bonded_atoms() method.
The maximum number of bonds an element can have is now taken into account so that protons are not
thought to be attached to 2 carbons within a 2 Angstrom radius!  The atoms closest to the atom of
interest are now connected.  Geometry constraints or other advanced techniques are not yet
implemented.