Clean up of the code before and after minfx execution.
This includes a number of simplifications.  pre_command_feed_back() is now called
initial_printout().  post_command_feedback() has been deleted.  This was for a kludge which can be
avoided by setting the minfx printout arg to zero.  The calls to set_generic_pre_and_post_amble()
and set_grid_pre_and_post_amble() were part of this kludge and are eliminated.  The self.silence
flag is now used to silence printouts when desired.