Expanded the print out when exiting to include references and shifted to using question().
This corresponds to one component of the patch_II file attached to https://gna.org/task/?6847 by
Michael Bieri (https://gna.org/users/michaelbieri).

The print out will need a bit of formatting, and it would be better if the question() box is
eventually removed as asking the user if they would like to quit after the user asked the program
to quit is annoying.