Changed text bottons to bitmap icons for the file opening in the model-free analysis GUI section.
This is mainly to test code manageability and design.  The '+' buttons have been replaced by oxygen
bitmap icons.  Note this is test code and can be reverted!  Issues found are as follows:

1)  The code for one GUI element is located in 3 places in the module, in 2 places in __init__()
(creation and event binding), and in __set_properties() to define sizes.  This makes GUI changes
difficult.

2)  The sizes in __set_properties() are hardcoded and repeated many, many times.  There should be
a variable defining x and y at the start for all these buttons.  These should be local variables of
a method containing the relaxation data input GUI element for model-free.

3)  I can't find how to modify the size of the box for the relaxation data input?!  This is visible
in the current revision.  The size of this box should be defined as the sum of the x and y lengths
of all the elements (not including a margin).