Abstraction of the images and icon paths.
The module 'paths' has been created to hold hardcoded variables pointing to the image and icon
paths.  All GUI modules now use this 'paths' module.  Therefore if an icon or image needs to be
changed, on change to the 'paths' module propagates to all GUI elements (avoids code hunting when
modifying paths).