Spun out the automatic model-free protocol code from the full_analysis.py script.
This is discussed in the post https://mail.gna.org/public/relax-devel/2010-01/msg00075.html.

The aim is to have a single code base for the automatic protocols for all UI designs (prompt,
script, GUI, web interface, etc.).  The API interface will be via class initialisations.  The
automatic model-free protocol has been placed into the dAuvergne_protocol class.

Future aims of this code is to have the UI fill a single data pipe with all the data required for
the analysis, and then to pass in this pipe into the protocol.  Then the auto-analysis will copy
this info into any other data pipes it requires.  This will allow for better GUI interaction, data
display, etc.