The relax base object is no longer passed into the class namespace of the user functions.
This was only being used for access to the intro flag (a few other cases exist, but these should be
fixed).  Now a new container object with this flag is being passed in.  This significantly
simplifies the code paths and removes the need for the test suite to create 'fake' relax instances.