The prompt info has been shifted into the execution info container.
The sys.ps1, sys.ps2, and sys.ps3 prompts are still set, but these strings are hardcoded into the
execution info container.  This allows the user functions to be imported and operate outside of the
prompt/scripting based interface, as the no longer require sys.ps3 to be set or the relax main
instance passed into them.