This patch fixed assignment of columns of parameter file.
This changed_parameterfile_settings_patch file was attached to https://gna.org/task/?6847 by Michael
Bieri (https://gna.org/users/michaelbieri), in the zip file calc_modelfree.py_patch.tar.gz
(https://gna.org/support/download.php?file_id=7750).

Note, this is only for testing the OMP data (https://gna.org/task/?6847#comment1). In future, this
can be set in a Settings Window (not included yet).

The patch was not fully applied, the RELAX_DATA variable was the only part changed.  The SEQ_DATA
change was whitespace at the end, which should not be there anyway.  The other changes are not of
use as the code was eliminated and replaced with a call to dAuvergne_protocol.