Added the initial code for data synchronisation between the GUI and the relax data store.
The state_load() and state_save() methods now call the new sync_ds() method of the main GUI class,
as well as sync_ds() methods for each analysis type.

The analysis specific synchronisation is performed as follows.  There will be a list in the main
class called self.analysis_frames.  This list will contain all the analysis frame objects open in
the GUI.  Each analysis type is expected to be a class instance, all located in
gui_bieri/auto_analyses/.  As these frames are initialised, they are appended to the main GUI
self.analysis_frames list.  Initialisation also creates a container in ds.relax_gui.analyses and
this is stored in the analysis class namespace as self.data.

Therefore if the model-free analysis is the 7th analysis opened, from within the main relax GUI
class the frame can be reached as:

self.analysis_frames[6]

Its method sync_ds() is reachable as:

self.analysis_frames[6].sync_ds()

The relax data store element for that analysis is reachable as:

ds.relax_gui.analyses[6]

or as an alias as:

self.analysis_frames[6].data

Within the frame instance namespace (in the class), the main GUI object is reachable as:

self.gui

This object oriented design will allow easy integration between the main window and the analysis
frames, however they may be laid out, and easy integration between analyses and the relax data
store.