The Auto_model_free.sync_ds() method is now complete.
Data can now be transfered both ways - to the relax data store and back again.  This enables the
program state to be saved and loaded (the ds.relax_gui.from_xml() method needs to be written).

A few changes were made for the frame info storage in the relax data store.  The 'structure_file'
variable has been added and the 'model_toggle' list of 10 booleans replaces the 'models' list.