Shifted the relax_control module down back into the gui_bieri package.
This is a GUI element that is independent of the specific analyses.  The analysis specific code can
be shifted into the modules of gui_bieri.analyses, and the threads also started from there.  This
window can be initialised in the main GUI class and then shown by the analyses' execute methods.