Shifted the peak list file and time info panel into add_peak_list_selection().
The code has been simplified by the renaming of all variables and most being shifted into the local
namespace.  The number of peak list elements in the panel can also be specified as an arg to
add_peak_list_selection().  The file names and times are now stored in lists to be looped over
rather than a separate variable for each.