Redesigned get_intro_string() to return a container of different strings.
The method is now called get_intro_text() and returns a container with the following str objects:
    - title:  The program title 'relax'
    - version:  For example 'repository checkout' or '1.3.8'.
    - desc:  The short program description.
    - copyright:  A list of copyright statements.
    - licence:  Text pertaining to the licencing.
    - errors:  A list of import errors.

This is being abstracted so that the different UIs can use this differently.  The prompt/script can
word wrap to 100 char (80 char in Windows), and print this to stdout.  The GUI can use this in the
about prompt and not print to stdout.  Other interfaces can use this in any other way.  And the
information content can be expanded for different purposes.