Bug fix for catching the looping around the universal solution in the dauvergne_protocol module.
This was first identified by Seb in the post at
https://mail.gna.org/public/relax-users/2007-09/msg00010.html (Message-id:
<46EEA6B2.90606@ulaval.ca>).

The problem was the automatic looping over optimisation cycles in the full_analysis.py script.  This
code is now in the dauvergne_protocol auto_analyses module.  The issue was a never-ending looping
around the universal solution (the optimisation minimum combined with Occam's razor or the model
selection minimum).  This should now be caught and the protocol terminated at the end of the first
completed loop.  The fix was to compare the chi2, selected models, diffusion tensor, and model-free
parameters to every iteration, starting from the first.  This will not be optimal if the protocol is
interrupted in the middle of one such loop, but will just mean that a few extra iterations will be
required to complete the loop.