Large clean up of the sequence information loading, either from a text file or from a PDB file.
Independent of the files selected, the NOE and Rx analyses will handle the file selections properly.
Proper synchronisation of the text fields in all analyses now occurs when the 'Molecule' menu
entries are used.  The sequence selection is now also shown in each frame above the PDB selection
and the text has been renamed to better reflect what those text inputs are for.