The diffusion tensor spheroid_type variable is now only updated if it is never explicitly set.
This is to fix the diffusion tensor setting bug found by Vitaly Vostri <vvostri att gmail dott com> 
(Message-id: <AANLkTinRO3ZCNwL6GeqWdx0uBvv=PQ=zw6vpdQQeruBW@mail.gmail.com>
https://mail.gna.org/public/relax-users/2010-12/msg00014.html).

The result of this bug is that initialising a prolate diffusion tensor with Da set to zero caused 
the spheroid_type to be automatically set to 'oblate'.  This bug was introduced at exactly r11678,
and was part of changes for better BMRB support.  This affects relax version 1.3.6 and the 1.3
repository line from r11678 to 12174.