Complete redesign of the relax global settings window to fix the Mac OS X crashing.
This is an attempt at fixing bug #18197 (https://gna.org/bugs/?18197).

The old code has been completely eliminated and replaced with a new set of methods in the base
relax_gui class and a new Global_parameters class in gui.settings which is of the same design as the
free file format settings window.