Modifying test_suite/shared_data/structures/Indol_test.xyz
This is Han's patch https://gna.org/task/download.php?file_id=13245 attached to the task
http://gna.org/task/?7404.

In the test_suite/shared_data/structures/Indol_test.xyz 0.00000 after the molecule name was removed.
The reason for that is because in a typical xyz file the second line is just the molecule name.