Debugging the user function load_spins() and add_item().
In the function load_spins() in generic_fns/structure/main.py a parameter 'res_num_exist' was added to check whether in the structural data the residue number exists or not. Normally in the pdb file there is a residue number and in the xyz file not. And then this parameter was used as a flag in the add_item() to avoid the Error message 'RelaxError("The residue number '" + repr(res_num) + "' already exists in the sequence.")' in case for xyz file when residue number does not exist.