Completely redesigned the results viewing code.
This is a huge code change!  The Results_summary class is now called Results_viewer is now a
wx.Frame.  The design is a non-modal dialog with three parts.  The top is the analysis selector
combo box.  The middle is the list of results files, updated to match the selected analysis.  The
bottom is a button for launching the appropriate viewer.  Most code is new, and a lot of old unused
code is still to be deleted.