Better error handling in the wizard.
The page objects now have a variable exec_status which is only set to True if the protected
execution returns true.  The wizard add_page() method now also has an argument called
proceed_on_error which allows the wizard pages to be configured to either remain showing when an
error occurs or to proceed to the next page.