Converted the NOE auto-analysis to operate on a pre setup data pipe.
The previous behaviour was to accept a series of arguments and then all the user functions were
executed within the auto-analysis.  However this eliminates a lot of the flexibility of relax.  So
now the user can set up their sequence data and peak height data as they see fit, and the
auto-analysis will take that as the starting point.