Replaced all calls to ds.__reset__() with calls to generic_fns.reset.reset().
This is needed as the relax data store is only part of the permanent data kept within relax.  The
status object and other data will need to be dealt with now.  The use of generic_fns.reset.reset()
by all parts of relax makes this resetting more uniform and clean.

The system and unit tests have have been simplified by shifting the tearDown() method into the
SystemTestCase and UnitTestCase base classes.  The tearDown() methods were mainly used to reset
relax, so this is now unified in the base class via generic_fns.reset.reset().