Converted the model-free auto-analysis tab to the new GUI design.
This now resembles the NOE and Rx GUI auto-analysis tabs in that the data pipe is created with a new
analysis and the data pipe will be passed into the auto-analysis.  The spin info is now set up using
the spin viewer window.  There is still a lot of work left to complete this auto-analysis.