Eliminated the global parameters data store object and GUI window.
This concept severely limits the GUI's flexibility and usefulness and is to be replaced by either
analysis specific GUI elements within the analysis pages, or by direct access to the corresponding
user function.  For example specifying 'newton' optimisation is useless for the R1 and R2 as only
simplex can be used, or the NOE as no optimisation is performed.