The notification of observers in status.observers.gui_uf has been modified.
The GUI user function pages no longer perform this as the calls to the interpreter are via the
interpreter thread, hence the completion of the page does not signal the completion of the user
function.  Instead the interpreter thread (and the synchronous apply() call as well) perform the
notification once complete.