Redesigned how the user function wizards are created.
The create_wizard() UF_base class method has been added to perform most of the work and to simplify
the specific user function code.  The __init__() method has also been reintroduced, but this time it
takes the 'parent' arg, which is the parent wx.Window, via the User_functions package class.