Added a toolbar button to the spin viewer window to launch a wizard for loading spins.
A simple wizard has been created for the spin viewer window for loading spins which directs the user
to either the sequence.read user function or the structure.read_pdb and structure.load_spins user
functions.  The wizard page Load_method_page has been added at the start to allow the user to choose
the method, and the last page Finish_page has been added so that the wizard can have a finish
button and terminal page.