Bug fix for the closure of the new analysis wizard window.
Closure of the wizard window would trigger an error in some of the wizards as the _status flag would
remain as True.  Now the close event is handled and the flag is set to False.  All other calls to
self.Close() have been replaced by Destroy() to bypass the closure event handler.