Bug fix for the GUI interpreter thread flushing on MS Windows.
The Queue.empty() method is documented as being unreliable - and that is indeed the case on MS
Windows (not on GNU/Linux or Mac OS X though).  It sometimes returns false, continually, when the
queue is completely empty!!!

Now the queue user function __repr__ strings are stored in a list to keep track of the queue status.