Fix for bug #18789 - the zero errors in Grace files.
This was submitted by Elio Cino <ecino att uwo dott ca> at https://gna.org/bugs/?18789.

The problem was that the errors were being taken as the X-axis error values, which in the case of
spin specific data is zero.  Now the Y-axis errors are taken rather than the unused X-axis errors.
This appears to have triggered a bug in the test suite relating to Grace files with both X and Y-
axis errors.