relax version 1.3.13.
This release introduces the second version of the relax GUI.  This is a major rewrite of the entire
GUI code base.  The GUI should now be much more flexible, being able to handle all the different
ways NMR data is collected and errors are measured.  It now has the full flexibility of the prompt/
scripting interface by the implementation of GUI versions of the user functions and now supports
small organic molecules, RNA, DNA, sugars, etc.  This flexibility will allow all other analysis
types (reduced spectral density mapping, N-state model, the frame order theory, consistency testing,
etc.) to be added to the GUI in the future.  The GUI is now fully tested and fully functional and
bug free on all major platforms including GNU/Linux, Mac OS X and MS Windows.

In addition, there are improvements to the frame order theory.  And significant bug fixes throughout
the program.