Converted the align_tensor.init user function to accept the align_id and domain arguments.
The domain argument complements the align_tensor.set_domain user function whereas the align_id arg
is new and is used to associate a tensor with an alignment.  This is part of a redesign of how
alignment tensors are handled, as there can be multiple tensors per alignment if the molecule is
multi-domain.