Created a special specific API object called SPIN_PARAMS.
This will be used to handle all operations to do with model parameters.  This is initialised in the
base class, and can be set up in the specific __init__() methods.  The object Param_list has methods
for parameter initialisation (where all info is specified such as Grace string, units, default
value, etc) and for determining if a parameter exists.  More methods will be added in the future to
expand the usability of this object and to simplify the specific API methods.