Mass conversion to the new GLOBAL_PARAMS and SPIN_PARAMS specific API data structures.
This involves mainly the conversion of the model-free analysis to these new objects, and the
subsequent changes to how model-free parameters are handled.  The parameters are now all lowercase,
for example ['S2', 'te', 'Rex'] is now ['s2', 'te', 'rex'].  The follow parameters are now converted
throughout relax:
    'bond_length' to 'r', 
    'CSA' to 'csa',
    'heteronucleus' to 'heteronuc_type',
    'proton' to 'proton_type'.

The regularisation of the specific analysis parameters via GLOBAL_PARAMS and SPIN_PARAMS has
significantly simplified the current code and make it much easier to introduce new analyses.