Redesigned the test data for the frame order models.
A base script for all models, 'generate_base.py', is now used to create all of the structural data
and to back-calculate the RDC and PCS data.  Rather than loading and rotating one structure per
state, the script now rotates the spin specific vectors and atomic positions.  This allows sampling
of up to 1e6 states to create very accurate debugging data.

All of the README files have been removed as these are repetitive.  The back_calc.py scripts and
save files have also been deleted as these are no longer used.  Two new files are created and have
been added to the repository.  The file 'axis_system' has a copy of the motional axis system and the
full precision Euler or spherical angles for the system.  The file 'frame_order_matrix' contains the
exact values of the frame order matrix and will be used for debugging.