Big fix for the IO streams when running in the multi-processor mpi4py mode!
This affects all multi-processor implementations, excluding the standard uni-processor.  The entire
stream capture design has been overhauled.  The IO streams of the master processor are no longer
touched, whereas those of the slaves are captured, stored in a list of lists where the original
'S  |' and 'S E|' tokens are prepended to each line as well as a number identifying the stream,
stored in the results command object, and unpacked and sent to sys.stdout and sys.stderr in the
correct order by the master.

This fixes the implementation for both running on the GUI (the slaves are no longer dumping text to
the terminal where the GUI was launched from), and on clusters where the IO is dumped on the
processing node rather than back on the master node.