Shifted the main interface of the multi-processor package into the __init__ module.
The two main interfaces, or the API with the calling program, are the load_multiprocessor() function
and the Processor_box class.  These have been shifted from the multi.processor module as this is
mainly a base class.  The import_module() function has also been shifted into __init__ and been made
a private function, removing it from the API.