Created a special Verbosity singleton for controlling the multi-processor package print outs.
This controls the text prepended to the output from the slaves, as well as all print outs from the
master processor.  If the verbosity level is set to 0, then nothing but warnings and errors are
printed.  The level of 1 prepends text to the slave IO streams and the print out of from the master
of the active slave set.

This will be important for very fine grained parallelisation whereby the multi-processor is used at
the level of the basic function call or lower.