Fix for bug #19528 (https://gna.org/bugs/?19528).
This bug was reported by Hugh Dannatt (h dott dannatt att sheffield dot ac dot uk).

The problem was in the multi-processor code for the model-free calculations, specifically for the
Monte Carlo simulations.  Rather than processing the queue of slave processor commands for the
model-free calculation for each spin system at the end of each simulation, the processing of the
queue was happening at the end of all Monte Carlo simulations.  This caused memory usage to bloat!
A full description of the problem can be found at
https://mail.gna.org/public/relax-users/2012-03/msg00037.html (Message-id:
<CAED9pY_Z1wGEg3gZo3WCkhvYTBdqs-Feex5Og4vJFovgJkUOyw@mail.gmail.com>).